+++ Commands/Receive-OBS.ps1
@@ -79,7 +79,7 @@ function Receive-OBS
         if ($PSCmdlet.ParameterSetName -eq 'WaitForResponse') {
             $myRequestId = $payloadObject.d.requestID
             $myRequestType = $payloadObject.d.requestType
-            if (-not $myRequestId -or -not $myRequestType) {
+            if (-not $myRequestId) {
                 Write-Error "No .RequestID to wait for"
                 return
             }
@@ -123,9 +123,19 @@ function Receive-OBS
                 }
     
                 # Otherwise, create a new PSObject out of the response
-                $responseObject = [PSObject]::new($responseObject)                    
-                # and decorate it with the command name and OBS.requestype.response
-                $responseObject.pstypenames.add("$myCmd")                        
+                $responseObject = [PSObject]::new($responseObject)
+                
+                # If the response is not a string
+                if ($responseObject -isnot [string]) {
+                    # clear the typename.
+                    $responseObject.pstypenames.clear()
+                }
+
+                if ($responseObject.inputKind) {
+                    $responseObject.pstypenames.add("OBS.Input.$($responseObject.inputKind -replace '_', '.')")
+                }
+                # Decorate the response with the command name and OBS.requestype.response
+                $responseObject.pstypenames.add("$myCmd")
                 $responseObject.pstypenames.add("OBS.$myRequestType.Response")
     
                 # Now, walk thru all properties in our input payload