+++ Formatting/OBS.GetInputList.Response.format.ps1
@@ -1,0 +1,1 @@
+Write-FormatView -TypeName 'OBS.GetInputList.Response' -Property InputName, InputKind