+++ Types/OBS.GetSceneItemList.Response/get_BlendMode.ps1
@@ -1,0 +1,1 @@
+$this.sceneItemBlendMode -replace '^OBS_BLEND_'