+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## obs-powershell 0.1.4:
+
+* Adding Add-OBSColorSource (Fixes #51)  
+* Save-OBSSourceScreenShot:
+  * Attaching .InputName, .SourceName, .ImageWidth, .ImageHeight to output (Fixes #50)
+  * Now returns a file (Fixes #49)
+* -Path parameters now attempt to resolve to an absolute path (Fixes #48)
+* All scene items can now:
+  * Blend() / get .BlendMode (Fixes #53)
+  * FitToScreen() (Fixes #46)
+  * Crop() (Fixes #57)
+  * Rotate() (Fixes #35)
+* Color Sources can now .SetColor (Fixes #55)
+
+---
+
 ## obs-powershell 0.1.3:
 
 * Requiring ThreadJob Module (Thanks @nyanhp!) (Fixes #36)